pyproject.toml
[tool.poetry]
name = "vadipartisweets"
version = "0.1.0"
description = ""
authors = ["Henish Patel <henishpatel9045@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
pillow = "^10.4.0"
djangorestframework = "^3.15.2"
pyautogui = "^0.9.54"
xlsxwriter = "^3.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
